+++ pydantic_ai_harness/__init__.py
@@ -4,9 +4,8 @@ from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
     from .code_mode import CodeMode
-    from .code_mode_dynamic_catalog import CodeModeDynamicCatalog
 
-__all__ = ['CodeMode', 'CodeModeDynamicCatalog']
+__all__ = ['CodeMode']
 
 
 def __getattr__(name: str) -> object:
@@ -14,8 +13,4 @@ def __getattr__(name: str) -> object:
         from .code_mode import CodeMode
 
         return CodeMode
-    if name == 'CodeModeDynamicCatalog':
-        from .code_mode_dynamic_catalog import CodeModeDynamicCatalog
-
-        return CodeModeDynamicCatalog
     raise AttributeError(f'module {__name__!r} has no attribute {name!r}')
