--- pydantic_ai_harness/code_mode_dynamic_catalog/__init__.py
@@ -1,5 +1,0 @@
-"""Cache-friendly tool catalog disclosure for CodeMode."""
-
-from ._capability import CodeModeDynamicCatalog
-
-__all__ = ['CodeModeDynamicCatalog']
